Attempt to fix race conditions

+++ Process/Instruction/Add.cpp
@@ -1,6 +1,6 @@
 #include "./Add.h"
 
-Add::Add(int var, int var1, int var2) : Instruction(ADD){
+Add::Add(uint16_t var, uint16_t var1, uint16_t var2) : Instruction(ADD){
     this->var=var;
     this->var1=var1;
     this->var2=var2;
